Add 'get_file_abspath'
The absolute path is the better identifier of files. With this help, we
can play file at it's folder with relative path and other folder with
absolute path, and we will get the record all the same.

P.S. another solution is using final file name, e.g., "2.mp4", this may
conflict with others.