Fix bug when finish new movie
If you play a new movie, which does not exist in json, and when you
finish it, it will execute remove method to santize json, but new file
has not been added to json, i.e., you try delete things not exist.

To solve this, add a validation before remove items in json.